+++ src/Tools/RandomShuffle.h
@@ -3,10 +3,10 @@
 
 #include "Random.h"
 
-static constexpr size_t best_prime[] = {2, 2, 5, 11, 17, 37, 67, 131, 257, 521, 1031, 2053, 4099, 8209, 16411, 32771, 65537, 131101, 262147, 524309, 1048583, 2097169, 4194319, 8388617};
+static constexpr uint32_t best_prime[] = {2, 2, 5, 11, 17, 37, 67, 131, 257, 521, 1031, 2053, 4099, 8209, 16411, 32771, 65537, 131101, 262147, 524309, 1048583, 2097169, 4194319, 8388617};
 
 //! Maximum number of elements for method 0
-static constexpr size_t shuffle_threshold = 8;
+static constexpr uint32_t shuffle_threshold = 8;
 
 // A random shuffler which ensures a small memory occupancy (fixed)
 // at the cost of a reasonable loss in performance
@@ -14,7 +14,7 @@ class RandomShuffle
 {
 public:
 
-    RandomShuffle( Random &rand, size_t length )
+    RandomShuffle( Random &rand, uint32_t length )
     : length_( length ), mask_( 1 ), P_( 0 ), i_( 0 )
     {
         reinit( rand, length );
@@ -22,7 +22,7 @@ public:
     
     //! Reinitialize the shuffler
     SMILEI_ACCELERATOR_DECLARE_ROUTINE
-    void reinit( Random &rand, size_t length ) {
+    void reinit( Random &rand, uint32_t length ) {
         i_ = 0;
         length_ = length;
         
@@ -30,12 +30,12 @@ public:
             
             // Fischer-Yates method
             
-            for( size_t i = 0; i < length; i++ ){
+            for( uint32_t i = 0; i < length; i++ ){
                 random_array_[i] = i;
             }
-            for( size_t i = length-1; i > 0; i-- ){
-                size_t j = rand.integer() % (i+1);
-                size_t a = random_array_[i], b = random_array_[j];
+            for( uint32_t i = length-1; i > 0; i-- ){
+                uint32_t j = rand.integer() % (i+1);
+                uint32_t a = random_array_[i], b = random_array_[j];
                 random_array_[i] = b; random_array_[j] = a;
             }
             
@@ -53,8 +53,8 @@ public:
             // Find the closest prime above 2^nbits
             P_ = best_prime[nbits];
             // Generate 7 random numbers that constitutes the key
-            for( size_t i=0; i<7; i++ ){
-                size_t r = rand.integer();
+            for( uint32_t i=0; i<7; i++ ){
+                uint32_t r = rand.integer();
                 // Prevent key = zero
                 do {
                     random_array_[i] = r & mask_;
@@ -66,7 +66,7 @@ public:
     SMILEI_ACCELERATOR_DECLARE_ROUTINE_END
     
     //! Get the next shuffled position
-    size_t next() {
+    uint32_t next() {
         if( length_ < shuffle_threshold ){
             return next0();
         } else {
@@ -75,30 +75,30 @@ public:
     }
     
     //! Get n next shuffled positions
-    void next( size_t n, size_t * shuffled ) {
+    void next( uint32_t n, uint32_t * shuffled ) {
         if( length_ < shuffle_threshold ){
             SMILEI_ACCELERATOR_LOOP_SEQ
-            for( size_t i = 0; i < n; i++ ) {
+            for( uint32_t i = 0; i < n; i++ ) {
                 shuffled[i] = next0();
             }
         } else {
             SMILEI_ACCELERATOR_LOOP_SEQ
-            for( size_t i = 0; i < n; i++ ) {
+            for( uint32_t i = 0; i < n; i++ ) {
                 shuffled[i] = next1();
             }
         }
     }
     
     //! Get the next shuffled position for method 0
-    size_t next0() {
-        size_t next_position = random_array_[i_];
+    uint32_t next0() {
+        uint32_t next_position = random_array_[i_];
         i_ = ( i_ + 1 ) % length_;
         return next_position;
     }
     
     //! Get the next shuffled position for method 1
-    size_t next1() {
-        size_t next_position;
+    uint32_t next1() {
+        uint32_t next_position;
         do{
             next_position = i_;
             i_ = ( i_ + 1 ) % mask_;
@@ -118,19 +118,19 @@ public:
 private:
     
     //! number of elements to permute
-    size_t length_;
+    uint32_t length_;
     
     //! A random array required initially for the algorithm
-    size_t random_array_[shuffle_threshold];
+    uint32_t random_array_[shuffle_threshold];
     
     //! Bitmask required for method 1
-    size_t mask_;
+    uint32_t mask_;
     
     //! Prime number required for method 1
-    size_t P_;
+    uint32_t P_;
     
     //! A number related to the number of calls to next()
-    size_t i_;
+    uint32_t i_;
     
 };
